# backend/nest/apiswagger.yaml
openapi: 3.0.1
info:
  title: FStudy - Language Course Purchase API
  description: |
    Welcome to the FStudy API! This API allows users to browse, purchase, and manage language courses on the FStudy platform. Our platform offers a wide range of language courses designed to help learners achieve their language learning goals efficiently and effectively.

    ## Features
    - **Course Catalog**: Browse a comprehensive list of available language courses.
    - **Course Details**: Get detailed information about each course, including syllabus, instructor information, and pricing.
    - **User Management**: Manage user profiles, including registration, login, and profile updates.
    - **Purchase Courses**: Securely purchase language courses and access purchased content.
    - **Progress Tracking**: Track learning progress and achievements within the platform.
    - **Reviews and Ratings**: Read and submit reviews and ratings for courses.

    ## Authentication
    The FStudy API uses JWT (JSON Web Token) for authentication. Users must obtain a token by logging in and include it in the `Authorization` header of requests.

    ## Error Handling
    The API returns standard HTTP status codes for successful and erroneous requests. Error responses include a message explaining the reason for the error.

    ## Base URL
    The base URL for all API requests is: `http://localhost:4000/`

    ## Versioning
...

    ## Contact
...

  contact:
    name: FStudy Support
...
  license:
...
servers:
...
paths:
...
  /course/type/create-course-type:
    post:
      tags:
      - CourseTypes
      summary: Create a new course type
      description: Create a new course type with a name and description.
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: "khóa học toeic 11"
                description:
                  type: string
                  example: "đây là khóa học tốt nhất 11"
      responses:
        '201':
          description: Course type created successfully
        '400':
          description: Invalid input
  /course/type/update-course-type/{id}:
    put:
      tags:
      - CourseTypes
      summary: Update an existing course type
      description: Update the name and description of an existing course type by its ID.
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: The course type ID
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: "Đay là tên khóa học sau khi chỉnh sửa mới nhất"
                description:
                  type: string
                  example: "Đây là mô tả khóa học sau khi chỉnh sửa mới nhất"
      responses:
        '200':
          description: Course type updated successfully
        '404':
          description: Course type not found
  /course/type/{id}:
    get:
      tags:
      - CourseTypes
      summary: Get a course type by ID
      description: Retrieve a specific course type by its ID.
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: The course type ID
      responses:
        '200':
          description: A specific course type
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    example: "6698a3084eab3a2333ffe388"
                  name:
                    type: string
                    example: "khóa học toeic"
                  description:
                    type: string
                    example: "đây là khóa học tốt nhất"
        '404':
          description: Course type not found
    delete:
      tags:
      - CourseTypes
      summary: Delete a course type by ID
      description: Delete a specific course type by its ID.
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: The course type ID
      responses:
        '204':
          description: Course type deleted successfully
        '404':
          description: Course type not found
...
  /course/content/create:
    post:
      tags:
      - Contents
      summary: Create a new content item
      description: Create a new content item with a specified type and associated content.
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                content_type:
                  type: string
                  example: "quiz"
                  description: The type of content to create. Examples include 'quiz', 'word_matching'.
                quiz:
                  type: string
                  example: "669b4e286fd08d69cfe87ae4"
                  description: The unique identifier for the quiz to be associated with this content.
      responses:
        '201':
          description: Content created successfully
        '400':
          description: Invalid input
  /course/content/{id}:
    get:
      tags:
      - Contents
      summary: Get content by ID
      description: Retrieve a specific content item by its ID.
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: The content item ID.
      responses:
        '200':
          description: A specific content item
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    example: "669a77d538cae0e5681fea1b"
                    description: The unique identifier for the content item.
                  content_type:
                    type: string
                    example: "quiz"
                    description: The type of content (e.g., quiz, word_matching).
                  quiz:
                    type: string
                    example: "669b4e286fd08d69cfe87ae4"
                    description: The unique identifier for the quiz associated with this content.
        '404':
          description: Content not found
    put:
      tags:
      - Contents
      summary: Update content
      description: Update the type and associated content of an existing content item by its ID.
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: The content item ID.
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                content_type:
                  type: string
                  example: "word_matching"
                  description: The updated type of content. Examples include 'quiz', 'word_matching'.
                word_matching:
                  type: string
                  example: "669a77d538cae0e5681fea1b"
                  description: The unique identifier for the word matching content to be associated with this content item.
      responses:
        '200':
          description: Content updated successfully
        '404':
          description: Content not found
    delete:
      tags:
      - Contents
      summary: Delete content by ID
      description: Delete a specific content item by its ID.
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: The content item ID.
      responses:
        '204':
          description: Content deleted successfully
        '404':
          description: Content not found
  /course/content/add/{id}:
    patch:
      tags:
      - Contents
      summary: Add content to an existing content item
      description: Add a specific type of content to an existing content item by its ID.
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: The content item ID.
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                contentType:
                  type: string
                  example: "quiz"
                  description: The type of content to add. Examples include 'quiz', 'word_matching'.
                dataId:
                  type: string
                  example: "669a0e8534614378b79d89d7"
                  description: The unique identifier for the additional content to be associated with this content item.
      responses:
        '200':
          description: Content added successfully
        '404':
          description: Content not found
...
  /course/lesson/create:
    post:
      tags:
      - Lessons
      summary: Create a new lesson
      description: Create a new lesson with specified content, title, and access type.
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                lesson:
                  type: string
                  example: "23"
                  description: The identifier or number of the lesson.
                content:
                  type: string
                  example: "669a63977976e3b25c8120d5"
                  description: The unique identifier for the content associated with the lesson.
                isFree:
                  type: boolean
                  example: true
                  description: Indicates if the lesson is free to access.
                title:
                  type: string
                  example: "Bài 1: bài cơ bản"
                  description: The title of the lesson.
      responses:
        '201':
          description: Lesson created successfully
        '400':
          description: Invalid input
  /course/lesson/{id}:
    get:
      tags:
      - Lessons
      summary: Get lesson by ID
      description: Retrieve a specific lesson by its ID.
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: The lesson ID.
      responses:
        '200':
          description: A specific lesson
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    example: "669c6d9713f60aad3988f9d0"
                    description: The unique identifier for the lesson.
                  title:
                    type: string
                    example: "Bài 1: bài cơ bản"
                    description: The title of the lesson.
                  content:
                    type: string
                    example: "669a63977976e3b25c8120d5"
                    description: The unique identifier for the content associated with the lesson.
                  isFree:
                    type: boolean
                    example: true
                    description: Indicates if the lesson is free to access.
        '404':
          description: Lesson not found
    put:
      tags:
      - Lessons
      summary: Update lesson
      description: Update the title, content, and access type of an existing lesson by its ID.
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: The lesson ID.
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                lesson:
                  type: string
                  example: "2"
                  description: The identifier or number of the lesson.
                content:
                  type: string
                  example: "669a0e8534614378b79d89d7"
                  description: The unique identifier for the content associated with the lesson.
                isFree:
                  type: boolean
                  example: true
                  description: Indicates if the lesson is free to access.
                title:
                  type: string
                  example: "Bài 1: bài nâng cao"
                  description: The title of the lesson.
      responses:
        '200':
          description: Lesson updated successfully
        '404':
          description: Lesson not found
    delete:
      tags:
      - Lessons
      summary: Delete lesson by ID
      description: Delete a specific lesson by its ID.
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: The lesson ID.
      responses:
        '204':
          description: Lesson deleted successfully
        '404':
          description: Lesson not found
  /course/lesson/add/{id}:
    patch:
      tags:
      - Lessons
      summary: Add content to a lesson
      description: Add content to an existing lesson by its ID.
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: The lesson ID.
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                content:
                  type: string
                  example: "669b4e3e6fd08d69cfe87ae6"
                  description: The unique identifier for the content to be added to the lesson.
      responses:
        '200':
          description: Content added to the lesson successfully
        '404':
          description: Lesson not found
  /course/quiz/create:
    post:
      tags:
      - Quizzes
      summary: Create a new quiz
      description: Create a new quiz with specified question, options, correct answer, and explanation.
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                question:
                  type: string
                  example: "Đâu là thủ đô của Việt Nam?"
                  description: The question to be answered in the quiz.
                options[0]:
                  type: string
                  example: "Thủ đô"
                  description: The first option for the answer.
                options[1]:
                  type: string
                  example: "Cần Thơ"
                  description: The second option for the answer.
                options[2]:
                  type: string
                  example: "Hà Nội"
                  description: The third option for the answer.
                options[3]:
                  type: string
                  example: "Hà Nội"
                  description: The fourth option for the answer.
                correctAnswer:
                  type: integer
                  example: 3
                  description: The index of the correct answer (starting from 0).
                explanation:
                  type: string
                  example: "Hà Nội is the capital of Vietnam."
                  description: An explanation for the correct answer.
      responses:
        '201':
          description: Quiz created successfully
        '400':
          description: Invalid input
  /course/quiz/update/{id}:
    put:
      tags:
      - Quizzes
      summary: Update quiz by ID
      description: Update the question, options, correct answer, and explanation of an existing quiz by its ID.
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: The quiz ID.
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                question:
                  type: string
                  example: "Đâu là thủ đô của Việt Nam vậy?"
                  description: The question to be answered in the quiz.
                options[0]:
                  type: string
                  example: "Cần Thơ"
                  description: The first option for the answer.
                options[1]:
                  type: string
                  example: "TP. Hồ Chí Minh"
                  description: The second option for the answer.
                options[2]:
                  type: string
                  example: "Hà Nội"
                  description: The third option for the answer.
                options[3]:
                  type: string
                  example: "Đồng Tháp"
                  description: The fourth option for the answer.
                correctAnswer:
                  type: integer
                  example: 2
                  description: The index of the correct answer (starting from 0).
                explanation:
                  type: string
                  example: "Hà Nội is the capital of Vietnam."
                  description: An explanation for the correct answer.
      responses:
        '200':
          description: Quiz updated successfully
        '404':
          description: Quiz not found
...
  /course/fill/create:
    post:
      tags:
      - Fill The Blank
      summary: Create a new fill-in-the-blank exercise
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                sentence:
                  type: string
                  description: "The sentence containing blanks, where each blank is represented by `____`."
                blanks:
                  type: array
                  items:
                    type: string
                  description: "An array containing the words or phrases that should fill in the blanks in the sentence."
                correctAnswers:
                  type: array
                  items:
                    type: string
                  description: "An array containing the correct answers for the blanks."
                correctAnswer:
                  type: string
                  description: "The overall correct answer for the exercise."
                  deprecated: true
                explanation:
                  type: string
                  description: "An explanation for the correct answers."
                  deprecated: true
      responses:
        '200':
          description: Exercise created successfully
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer 
      bearerFormat: JWT